refactor drag utils
closes #326
