+++ projects/ngx-d3timeline/src/lib/activity/activity.selectors.ts
@@ -5,22 +5,21 @@ import {
 } from '../store/state';
 import { selectTimeOrientation } from '../options/selectors/options.selectors';
 import { createSelector } from '../store-lib/selector/create-selector';
-import {
-  findActivity,
-  getDragPointInResourceAxis,
-  valueToResource,
-  getNonDraggedActivities,
-  getDragEventId,
-  getCurrentlyDraggedActivityWithDraggedToResource,
-  getDeltaTimeForDrag,
-  updateActivityForDrag,
-  shiftTimeByRangeValue
-} from '../drag/drag-utils';
+import { findActivity } from '../drag/drag-utils';
+import { TimelineDragEvent } from '../drag/timeline-drag-event';
+import { identifier } from '../core/types';
+import { PositionedActivity } from './positioned-activity';
+import { InverseBandScale, TimeScale } from '../scales/scale-types';
+import { Orientation } from '../core/orientation';
 import { selectBandScale, selectTimeScale } from '../scales/scale-selectors';
 import { getInverseBandScale } from '../scales/scale-utils';
 import { MemoizedSelector } from '../store-lib/selector/memoized-selector';
 
 const selectDragEventId = createSelector(selectDragEvent, getDragEventId);
+
+function getDragEventId(dragEvent: TimelineDragEvent): identifier {
+  return dragEvent && dragEvent.id;
+}
 
 export const selectNonDraggedActivities = createSelector(
   selectPositionedActivities,
@@ -28,11 +27,32 @@ export const selectNonDraggedActivities = createSelector(
   getNonDraggedActivities
 );
 
+function getNonDraggedActivities(
+  activities: PositionedActivity[],
+  dragEventId: identifier
+): PositionedActivity[] {
+  return dragEventId
+    ? activities.filter(activity => activity.id !== dragEventId)
+    : activities;
+}
+
 export const selectCurrentlyDraggedActivity = createSelector(
   selectPositionedActivities,
   selectDragEventId,
   findActivity
 );
+
+function getCurrentlyDraggedActivityWithDraggedToResource(
+  currentlyDraggedActivity: PositionedActivity,
+  updatedResource: string
+): PositionedActivity {
+  return (
+    currentlyDraggedActivity && {
+      ...currentlyDraggedActivity,
+      updatedResource
+    }
+  );
+}
 
 const selectInverseBandScale = createSelector(
   selectBandScale,
@@ -45,17 +65,44 @@ const selectDragPointInResourceAxis = createSelector(
   getDragPointInResourceAxis
 );
 
+function getDragPointInResourceAxis(
+  timeOrientation: Orientation,
+  dragEvent: TimelineDragEvent
+): number {
+  return (
+    dragEvent &&
+    (timeOrientation === Orientation.Vertical ? dragEvent.x : dragEvent.y)
+  );
+}
+
 const selectDraggedToResource = createSelector(
   selectDragPointInResourceAxis,
   selectInverseBandScale,
   valueToResource
 );
 
+export function valueToResource(
+  value: number,
+  inverseBandScale: InverseBandScale
+): string {
+  return inverseBandScale(value);
+}
+
 const selectDeltaTimeForDrag = createSelector(
   selectTimeOrientation,
   selectDragEvent,
   getDeltaTimeForDrag
 );
+
+function getDeltaTimeForDrag(
+  timeOrientation: Orientation,
+  dragEvent: TimelineDragEvent
+): number {
+  return (
+    dragEvent &&
+    (timeOrientation === Orientation.Vertical ? dragEvent.dy : dragEvent.dx)
+  );
+}
 
 const selectUpdatedStartForCurrentlyDraggedActivity = createSelector(
   selectCurrentlyDraggedActivity,
@@ -75,6 +122,14 @@ const selectTimeShiftedForDragEvent = (selectTime: MemoizedSelector<Date>) =>
     shiftTimeByRangeValue
   );
 
+function shiftTimeByRangeValue(
+  time: Date,
+  timeScale: TimeScale,
+  rangeValue: number
+): Date {
+  return timeScale.invert(timeScale(time) + rangeValue);
+}
+
 export const selectCurrentlyDraggedActivityWithDraggedToResource = createSelector(
   selectCurrentlyDraggedActivity,
   selectDraggedToResource,
@@ -88,6 +143,22 @@ export const selectActivityUpdatedForDrag = createSelector(
   selectDraggedToResource,
   updateActivityForDrag
 );
+
+function updateActivityForDrag(
+  positionedActivity: PositionedActivity,
+  updatedStart: Date,
+  updatedFinish: Date,
+  updatedResource: string
+): PositionedActivity {
+  return (
+    positionedActivity && {
+      ...positionedActivity,
+      updatedStart,
+      updatedFinish,
+      updatedResource
+    }
+  );
+}
 
 export const selectLastDraggedActivity = createSelector(
   selectPositionedActivities,
